requirements.txt
acme==0.31.0
asn1crypto==0.24.0
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==2.2
django-geojson==2.12.0
django-leaflet==0.24.0
future==0.15.2
idna==2.6
josepy==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
mock==2.0.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
parsedatetime==2.4
pbr==3.1.1
pyasn1==0.4.2
pycrypto==2.6.1
pygobject==3.26.1
PyICU==1.9.8
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.8.0
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.3.0
ssh-import-id==5.7
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
uWSGI==2.0.18
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
